Add licences, move code
